Keep camera frames as uint8 instead of widening them to float32
The frames are 8 bit at the sensor, so widening them bought nothing: it cost
2.74 ms and 2.8 MB of extra memory traffic per 640x480 frame and quadrupled
every recording. Cameras now hand out uint8 CHW frames (0-255) and depth stays
uint16 millimetres, which is the z16 format the RealSense already reports.

  640x480 frame     before        after
  post-processing   2.74 ms       0.32 ms
  bytes             3,686,400     921,600

At 30 fps across two cameras that is 164 ms/s of CPU down to 19 ms/s.

This is a recorded-data format change:

* RealsenseCamera and WebCamera return NDArray[np.uint8]; read_depth() and
  async_read_depth() return NDArray[np.uint16] shaped (1, H, W).
* H5Handler now preserves dtypes on save and load. It previously forced
  float32 both ways, which would have thrown the saving away. BLOSC already
  preserved them. Plain Python lists still default to float32, so callers
  passing lists are unaffected.
* Tactile and audio data stay float32, and so do arm observations. The dicts
  that hold colour and depth side by side are typed as the union of the two.
* Normalisation is deliberately left to the consumer: 0..1 versus ImageNet
  statistics is a per-model choice, and doing it here forces the most
  expensive representation on the recording path too.

Recordings made before this change stay float32 and still load; code mixing
old and new data should branch on dtype.

The shared post-processing moved into sensors/visual/frame_ops.py, which both
cameras now use. That removed a latent bug on the way: the "unexpected frame
shape" rescue in RealsenseCamera.read()/async_read() would transpose(2, 1, 0)
a single-channel CHW frame into nonsense; the conversion helper guarantees CHW
so the rescue is gone. The animation preprocessing in RakudaSaveWorker also
had to change: uint8 / 255.0 promotes to float64, which would have made a
150-frame recording 1.4 GB in RAM. It now hands uint8 to imshow directly
(matplotlib reads it as 0..255) and normalises anything else in float32.

Testing: tests/test_camera_frames.py covers the frame contract, the HDF5
round trip and the animation preprocessing, none of which needs a camera.
The CI workflow (renamed, since it is no longer Dynamixel-only) runs them.

+++ src/robopy/config/robot_config/rakuda_config.py
@@ -60,7 +60,7 @@ class RakudaArmObs:
 
 @dataclass
 class RakudaSensorObs:
-    cameras: Dict[str, NDArray[np.float32] | None]
+    cameras: Dict[str, NDArray[np.uint8] | None]
     tactile: Dict[str, NDArray[np.float32] | None]
     audio: Dict[str, NDArray[np.float32] | None]
