+++ src/robopy/config/robot_config/xarm_config.py
@@ -243,7 +243,7 @@ class XArmArmObs:
 
 @dataclass
 class XArmSensorObs:
-    cameras: Dict[str, NDArray[np.float32] | None]
+    cameras: Dict[str, NDArray[np.uint8] | None]
     tactile: Dict[str, NDArray[np.float32] | None]
     audio: Dict[str, NDArray[np.float32] | None]
