+++ src/robopy/robots/so101/so101_spacemouse.py
@@ -241,7 +241,7 @@ class So101SpaceMouseController:
         # Main thread: capture sensor data at *fps*
         leader_obs: List[NDArray[np.float32]] = []
         follower_obs: List[NDArray[np.float32]] = []
-        camera_obs: DefaultDict[str, List[NDArray[np.float32] | NDArray[np.uint8] | None]] = (
+        camera_obs: DefaultDict[str, List[NDArray[np.uint8] | NDArray[np.uint16] | None]] = (
             defaultdict(list)
         )
 
@@ -329,7 +329,7 @@ class So101SpaceMouseController:
         follower_arr = np.asarray(follower_obs, dtype=np.float32)
         arms = So101ArmObs(leader=leader_arr, follower=follower_arr)
 
-        camera_obs_np: Dict[str, NDArray[np.uint8] | NDArray[np.float32] | None] = {}
+        camera_obs_np: Dict[str, NDArray[np.uint8] | NDArray[np.uint16] | None] = {}
         for cam_name, frames in camera_obs.items():
             if frames and all(f is not None for f in frames):
                 camera_obs_np[cam_name] = np.asarray(frames)
